manage ad ad usertogroup works

+++ ManageAd.ps1
@@ -1,5 +1,6 @@
 . .\MySQL.ps1
 . .\GlobalVariables.ps1
+. .\Logger.ps1
 
 function deactivateDeletedStudents {
     param (
@@ -12,7 +13,11 @@ function deactivateDeletedStudents {
         $adName = $adStudent.Name
         [bool]$isUserInADAndInDB = 0
         foreach($dbStudent in $dbStudents ){
-            $dbStudentMax20 = $dbStudent[1].subString(0,20)
+            $dbStudentMax20 = $dbStudent[1]      
+            if ($dbStudentMax20.length -gt 20) {
+                $dbStudentMax20 = $dbStudent[1].subString(0,20)
+            }
+            
             if ($adName -eq $dbStudentMax20) {
                 $isUserInADAndInDB = 1
             }
@@ -51,40 +56,50 @@ function addUserToGroup {
     param (
     )
             $querry = "SELECT schueler.username, klasse.klassenbezeichnung FROM schueler
-            INNER JOIN `schueler-klasse` ON schueler.sid = `schueler-klasse`.sid
-            INNER JOIN klasse ON klasse.kid = `schueler-klasse`.kid"
+            INNER JOIN ``schueler-klasse`` ON schueler.sid = ``schueler-klasse``.sid
+            INNER JOIN klasse ON klasse.kid = ``schueler-klasse``.kid"
             $dbStudentGroupNameAssigne = runSql $GlobalDatabaseName $querry
             foreach($assigne in $dbStudentGroupNameAssigne)
             {
-                $adUser = GET-ADUser -Identity $assigne[0] -SearchScope $GlobalStudentOUPath
-                Get-ADGroup -Identity $assigne[1] -SearchScope $GlobalGroupOUPath | Add-ADGroupMember $adUser
+                $groupname= "GISO_$($assigne[1])"
+                $adUser = GET-ADUser -Filter "Name -eq '$($assigne[0])'" -SearchBase $GlobalStudentOUPath                
+                try {
+                    Get-ADGroup -Filter "Name -eq '$groupname'" -SearchBase $GlobalGroupOUPath | Add-ADGroupMember -Members $adUser
+                    log "user added to group, user: $($adUser.name), group: $groupname" "INFO"
+                }
+                catch {
+                    log "cant add user to Group $($adUser.name), group: $groupname, message: $_"
+                }
             }
 }
 
 function removeOldAdGroupsFromStudent {
     param (
     )
-    $querry = "SELECT schueler.username, klasse.klassenbezeichnung FROM schueler
-    INNER JOIN `schueler-klasse` ON schueler.sid = `schueler-klasse`.sid
-    INNER JOIN klasse ON klasse.kid = `schueler-klasse`.kid"
-    $dbStudentGroupAssigne = runSql $GlobalDatabaseName $querry
     $dbStudents = runSql $GlobalDatabaseName "SELECT username FROM schueler"
     foreach($dbStudent in $dbStudents){
-        $currentlyGroupMember = Get-ADPrincipalGroupMembership -Identity $dbStudent[1]
+        $currentlyGroupMember = Get-ADPrincipalGroupMembership -Identity $dbStudent[0]
         [bool]$isAdGroupInDB = $false
         foreach($group in $currentlyGroupMember)
         {
-            foreach($dbGroup in ($dbStudentGroupAssigne | Where-Object $_[0] -eq $dbStudent[1]))
-            {            
-                if ($group.name -eq $dbGroup[1]) {
-                    $isAdGroupInDB = $true
+            if ($group.DistinguishedName -like "CN=$groupname, $GlobalGroupOUPath") {
+                $querry = "SELECT schueler.username, klasse.klassenbezeichnung FROM schueler
+                INNER JOIN ``schueler-klasse`` ON schueler.sid = ``schueler-klasse``.sid
+                INNER JOIN klasse ON klasse.kid = ``schueler-klasse``.kid
+                WHERE schueler.username = '$($dbStudent[0])'"
+                $dbStudentGroupAssigne = runSql $GlobalDatabaseName $querry
+                foreach($dbGroup in $dbStudentGroupAssigne)
+                {   
+                    $dbToAdGroupName = "GISO_$($dbGroup[1])"         
+                    if ($group.name -eq $dbToAdGroupName) {
+                        $isAdGroupInDB = $true
+                    }
+                }
+                if (!$isAdGroupInDB) {
+                    $adMember = Get-ADUser -Identity $dbStudent[0]
+                    $group | Remove-ADGroupMember -Members $adMember
                 }
             }
-
-        }
-        if (!$isAdGroupInDB) {
-            $adMember = Get-ADUser -Identity $dbStudent[1]
-            $group | Remove-ADGroupMember -Members $adMember
         }
     }
 }
@@ -103,4 +118,5 @@ function removeEmptyGroups {
 
 deleteDeletedGroups
 deactivateDeletedStudents
+addUserToGroup
 removeOldAdGroupsFromStudent
